feat: Implement order management functionality

+++ src/main/java/com/shirtshop/entity/Product.java
@@ -13,7 +13,7 @@ public class Product {
     @Id private String id;
     private String name;
     private String description;
-    private BigDecimal price;
+    private int price;
     private String category;
     private List<String> imageUrls;
     private List<String> imagePublicIds;
